chore: scaffolding for integration test
